add badge
