+++ pages/index.tsx
@@ -2,7 +2,6 @@ import { containerRef } from "@/components/Layout";
 import { Navbar } from "@/components/Navbar";
 import { AvailabilityTrigger } from "@/components/Start/AvailabilityTrigger";
 import { Friend } from "@/components/Start/Friend";
-import { SearchFriend } from "@/components/Start/SearchFriend";
 import { StatusSelector } from "@/components/Start/StatusSelector";
 import { useSession } from "@/lib/client/session";
 import { useColor, useDarkMode } from "@/lib/client/useColor";
@@ -14,6 +13,7 @@ import { useMemo } from "react";
 import { Virtuoso } from "react-virtuoso";
 import useSWR from "swr";
 import { HeadingComponent } from "../components/Start/HeadingComponent";
+import { FriendsTrigger } from "@/components/Start/FriendsTrigger";
 const ContactSync = dynamic(() => import("@/components/Start/ContactSync"));
 
 export default function Home() {
@@ -98,7 +98,7 @@ export default function Home() {
         >
           <AvailabilityTrigger />
           <StatusSelector mutate={mutate} profile={profileData} />
-          <SearchFriend mutate={mutate} />
+          <FriendsTrigger />
         </Box>
         <Box
           sx={{
